separate out workspace access and "auth" functionality
move to independent provider to prevent unnecessary fetches before a token has been provided
